Brief Insight on the Project:
Devices used: Particle Photon Device and an OLED Display Board. 
Code Description: Obtains the CO2 levels, humidity, and temperature from the sensors and displays it on the OLED display. It publishes the data on the Particle Console and graphs it on an IoT Analytics platform called Thingspeak, as an integrated Particle Webhook.